+++ src/moonblade/functions/strings.rs
@@ -152,3 +152,26 @@ pub fn replace(args: BoundArguments) -> FunctionResult {
 
     Ok(DynamicValue::from(replaced))
 }
+
+pub fn chr(mut args: BoundArguments) -> FunctionResult {
+    let n = args.pop1_number()?;
+
+    let u32c: u32 = n
+        .as_int()
+        .try_into()
+        .map_err(|_| "given number does not fit a u32")?;
+
+    let c = char::from_u32(u32c).ok_or("given number is out-of-range")?;
+
+    Ok(c.into())
+}
+
+pub fn ord(args: BoundArguments) -> FunctionResult {
+    let string = args.get1_str()?;
+    let mut chars = string.chars();
+
+    match (chars.next(), chars.next()) {
+        (Some(c), None) => Ok((c as i64).into()),
+        _ => Err("expected a single character")?,
+    }
+}